requirements.txt
asgiref==3.6.0
black==23.3.0
bulma==1.0.1
click==8.1.3
dj-inmemorystorage==2.1.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==23.5
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pytz==2022.7
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
